% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExprGenes.R
\name{diffExprGenes}
\alias{diffExprGenes}
\title{3) Differentially Expressed Genes (DEGs)}
\usage{
diffExprGenes(norm, norm_plots, cons, projectId, workspace)
}
\arguments{
\item{norm}{ExpressionFeatureSet object with normalized data from QCnorm function}

\item{norm_plots}{ExpressionFeatureSet object with normalized and ComBat batch corrected data from QCnorm function}

\item{cons}{List groups to compare, groups must match assignments in listGroups param of previous functions}

\item{projectId}{A unique identifier for the project}

\item{workspace}{Working directory}
}
\value{
List of DEGs for each contrast, annotated normalized data, and pheno data
}
\description{
3) Differentially Expressed Genes (DEGs)
}
\note{
Baseline/control group should be listed second in contrast assignment: c("experimentalGroup-ControlGroup","diseasedGroup-ControlGroup")

Outputs volcano plot (using plotly) of differentially expressed genes
}
\examples{
diff_expr_genes = diffExprGenes(norm_celfiles,c("RNA_1-Ctl"),'NCI_Project_1','/Users/name/folderName')
diff_expr_genes = diffExprGenes(norm_celfiles,c("KO_1-Ctl_1"),'NCI_Project_2','/Users/name/folderName')
}
\references{
See plotly, limma and Bioconductor platform design/annotation documentation
}
