% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneHeatmap.R
\name{geneHeatmap}
\alias{geneHeatmap}
\title{5) Gene heatmap with selected genes from l2p pathways}
\usage{
geneHeatmap(degs, paths, contrast, upOrDown, pathway_name,
  saveImageFileName)
}
\arguments{
\item{degs}{Differentially expressed genes object from deg function}

\item{paths}{Pathways object from pathways function}

\item{contrast}{User-selected contrast to display, must be a contrast from listContrasts argument of deg function}

\item{upOrDown}{Is user-selected pathway from up or down list? (matches names of up and down pathways from output of pathways function)}

\item{pathway_name}{What is the pathway name? Found in description column of pathway table}

\item{saveImageFileName}{Name of heatmap}
}
\description{
5) Gene heatmap with selected genes from l2p pathways
}
\note{
Nothing to return, outputs heatmap
}
\examples{
geneHeatmap(diff_expr_genes, l2p_pathways, 'RNA_1-Ctl', 'upregulated_pathways','oxidation-reduction process','Heatmap_Redox') 
geneHeatmap(diff_expr_genes, l2p_pathways, 'KO_1-Ctl_1', 'downregulated_pathways','G-protein coupled receptor activity','Heatmap_GPCR') 
}
\references{
See pheatmap package, mouse/human homologs extracted from http://www.informatics.jax.org/homology.shtml
}
