% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocalGEOfiles.R
\name{getLocalGEOfiles}
\alias{getLocalGEOfiles}
\title{1) Process GEO mouse or human data from an experiment with raw Affymetrix CEL files, see https://www.ncbi.nlm.nih.gov/geo/}
\usage{
getLocalGEOfiles(projectId, id, listGroups, listBatches = NULL, workspace)
}
\arguments{
\item{projectId}{A unique identifier for the project}

\item{id}{A GSE id number}

\item{listGroups}{Group assignments for each sample, follow order of samples on GEO series page}

\item{listBatches}{Optional list of batches for each sample, follow alphabetical order of samples}

\item{workspace}{Working directory}
}
\value{
ExpressionFeatureSet object with raw data and phenotype information
}
\description{
1) Process GEO mouse or human data from an experiment with raw Affymetrix CEL files, see https://www.ncbi.nlm.nih.gov/geo/
}
\examples{
celfiles = getLocalGEOfiles(projectId='NCI_Project_1',id='GSE61989',listGroups=c('Ctl','Ctl','Ctl','Ctl','RNA_1','RNA_1','RNA_1','RNA_1','RNA_2','RNA_2','RNA_2','RNA_2'),workspace='/Users/name/folderName')      
celfiles = getLocalGEOfiles(projectId='NCI_Project_2',id='GSE106988',listGroups=c('Ctl_1','Ctl_1','Ctl_1','KO_1','KO_1','KO_1'),listBatches=c(rep('A',6),rep('B',6)),workspace='/Users/name/folderName')
}
\references{
See packages oligo, GEOquery, Biobase
}
