% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2pPathways.R
\name{l2pPathways}
\alias{l2pPathways}
\title{l2p Pathway Analysis}
\usage{
l2pPathways(degs, species, workspace, projectId, configuration_path)
}
\arguments{
\item{degs}{Differentially expressed genes object from deg function}

\item{species}{Species of organism (human or mouse)}

\item{workspace}{Working directory}

\item{projectId}{A unique identifier for the project}

\item{configuration_path}{Path to configuration files}
}
\value{
List of up and downregulated pathways for each contrast
}
\description{
l2p Pathway Analysis
}
\examples{
l2p_pathways = l2pPathways(diff_expr_genes,'human','/Users/name/folderName','NCI_Project_1','/Users/name/config')
l2p_pathways = l2pPathways(diff_expr_genes,'mouse','/Users/name/folderName','NCI_Project_2','/Users/name/config')
}
\references{
l2p and m2h from CCBR/CCR/NCI/NIH
}
