% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{pathways}
\alias{pathways}
\title{l2p Pathway Analysis}
\usage{
pathways(degs, species, workspace, projectId)
}
\arguments{
\item{degs}{Differentially expressed genes object from deg function}

\item{species}{Species of organism (human or mouse)}

\item{workspace}{Working directory}

\item{projectId}{A unique identifier for the project}
}
\value{
List of up and downregulated pathways for each contrast
}
\description{
l2p Pathway Analysis
}
\examples{
l2p_pathways = pathways(diff_expr_genes,'human','/Users/name/folderName','NCI_Project_1')
l2p_pathways = pathways(diff_expr_genes,'mouse','/Users/name/folderName','NCI_Project_2')
}
\references{
l2p and m2h from CCBR/CCR/NCI/NIH
}
