% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCnorm.R
\name{QCnorm}
\alias{QCnorm}
\title{2) Runs QC analyses and normalizes data.}
\usage{
QCnorm(raw, path)
}
\arguments{
\item{raw}{ExpressionFeatureSet from either processGEOfiles or processCELfiles function}

\item{path}{Directory for plots}
}
\value{
Normalized ExpressionFeatureSet object with normalized data and phenotype information
}
\description{
2) Runs QC analyses and normalizes data.
}
\note{
Normalizes using rma, see oligo package

Outputs 3D PCA and similarity heatmap

Outputs pre-normalization plots, QC plots, post-normalization plots, see oligo package for details
}
\examples{
norm_celfiles = QCnorm(celfiles,'Users/name/folderName/plots') 
}
\references{
See packages rgl, Biobase, heatmaply, oligo
}
