% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMA_QCnorm.R
\name{RMA_QCnorm}
\alias{RMA_QCnorm}
\title{2) Runs QC analyses and normalizes data.}
\usage{
RMA_QCnorm(raw, path, contrast, listBatches = NULL)
}
\arguments{
\item{raw}{ExpressionFeatureSet from either processGEOfiles or processCELfiles function}

\item{path}{Directory for plots}

\item{contrast}{Contrast for subsetting QC plots by samples in chosen groups}

\item{listBatches}{Optional list of batches for each sample, follow alphabetical order of samples}
}
\value{
Normalized ExpressionFeatureSet object with normalized data and phenotype information
}
\description{
2) Runs QC analyses and normalizes data.
}
\note{
Normalizes using rma, see oligo package

Outputs 3D PCA and similarity heatmap

Outputs pre-normalization plots, QC plots, post-normalization plots, see oligo package for details
}
\examples{
norm_celfiles = RMA_QCnorm(celfiles,'Users/name/folderName/plots',c("RNA_1-Ctl")) 
}
\references{
See packages rgl, Biobase, heatmaply, oligo
}
