% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgseaPathways.R
\name{ssgseaPathways}
\alias{ssgseaPathways}
\title{6) ssGSEA function}
\usage{
ssgseaPathways(deg_normAnnot, species, geneSet, workspace, projectId,
  configuration_path)
}
\arguments{
\item{deg_normAnnot}{Differentially expressed genes object from deg function}

\item{species}{Species of organism (human or mouse)}

\item{geneSet}{Choice of MSigDB gene set}

\item{workspace}{Working directory}

\item{projectId}{A unique identifier for the project}

\item{configuration_path}{Path to configuration files}
}
\value{
List containing ssGSEA initial results and differentially expressed ssGSEA pathways
}
\description{
6) ssGSEA function
}
\note{
Human Gene Set Choices: "H: Hallmark Gene Sets", "C1: Positional Gene Sets", "C2: Curated Gene Sets", "C3: Motif Gene Sets", "C4: Computational Gene Sets","C5: GO gene sets", "C6: Oncogenic Signatures", "C7: Immunologic Signatures"

Mouse Gene Set Choices: "H: Hallmark Gene Sets", "C2: Curated Gene Sets", "C3: Motif Gene Sets", "C4: Computational Gene Sets", "C5: GO gene sets", "C6: Oncogenic Signatures", "C7: Immunologic Signatures"
}
\examples{
ssGSEA_results = ssgseaPathways(diff_expr_genes,'human','C2: Curated Gene Sets','/Users/name/folderName','NCI_Project_1')
}
\references{
MSigDB: http://software.broadinstitute.org/gsea/msigdb/index.jsp.  Also see GSEABase, GSVA and pheatmap packages.

Mouse gene sets adapted from http://bioinf.wehi.edu.au/software/MSigDB/
}
