% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCELfiles.R
\name{getCELfiles}
\alias{getCELfiles}
\title{1) Process local Affymetrix CEL files for mouse or human data}
\usage{
getCELfiles(projectId, listGroups, workspace)
}
\arguments{
\item{projectId}{A unique identifier for the project}

\item{listGroups}{Group assignments for each sample, follow alphabetical order of samples}

\item{workspace}{Working directory}
}
\value{
ExpressionFeatureSet object with raw data and phenotype information
}
\description{
1) Process local Affymetrix CEL files for mouse or human data
}
\examples{
celfiles = processCELfiles('NCI_Project_1',c('Ctl_1','Ctl_1','Ctl_1','KO_1','KO_1','KO_1'),'/Users/name/folderName')
celfiles = processCELfiles('NCI_Project_2',c('Ctl','Ctl','Ctl','Ctl','RNA_1','RNA_1','RNA_1','RNA_1','RNA_2','RNA_2','RNA_2','RNA_2'),'/Users/name/folderName')      
}
\references{
See packages tools, Biobase, oligo
}
