% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocalGEOfiles.R
\name{getLocalGEOfiles}
\alias{getLocalGEOfiles}
\title{1) Process GEO mouse or human data from an experiment with raw Affymetrix CEL files, see https://www.ncbi.nlm.nih.gov/geo/}
\usage{
getLocalGEOfiles(projectId, id, listGroups, workspace)
}
\arguments{
\item{projectId}{A unique identifier for the project}

\item{id}{A GSE id number}

\item{listGroups}{Group assignments for each sample, follow order of samples on GEO series page}

\item{workspace}{Working directory}
}
\value{
ExpressionFeatureSet object with raw data and phenotype information
}
\description{
1) Process GEO mouse or human data from an experiment with raw Affymetrix CEL files, see https://www.ncbi.nlm.nih.gov/geo/
}
\examples{
celfiles = processGEOfiles('NCI_Project_1','GSE61989',c('Ctl','Ctl','Ctl','Ctl','RNA_1','RNA_1','RNA_1','RNA_1','RNA_2','RNA_2','RNA_2','RNA_2'),'/Users/name/folderName')      
celfiles = processGEOfiles('NCI_Project_2','GSE106988', c('Ctl_1','Ctl_1','Ctl_1','KO_1','KO_1','KO_1'),'/Users/name/folderName')
}
\references{
See packages oligo, GEOquery, Biobase
}
